requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
seaborn
plotly
Pillow
